+++ lib/src/haversine_base.dart
@@ -14,14 +14,35 @@ class Haversine {
 
         this.latitude2 = _radiansFromDegrees(latitude2);
         this.longitude2 = _radiansFromDegrees(longitude2);
+
+        _throwExceptionOnInvalidCoordinates();
     }
 
-    Haversine.fromRadiams({this.latitude1, this.longitude1, this.latitude2, this.longitude2}) {
+    Haversine.fromRadians({this.latitude1, this.longitude1, this.latitude2, this.longitude2}) {
         this.latitude1 = latitude1;
         this.longitude1 = longitude1;
 
         this.latitude2 = latitude2;
         this.longitude2 = longitude2;
+
+        _throwExceptionOnInvalidCoordinates();
+    }
+
+    void _throwExceptionOnInvalidCoordinates() {
+
+        String invalidDescription = """
+            A coordinate is considered invalid if it meets at least one of the following criteria:
+            - Its latitude is greater than 90 degrees or less than -90 degrees.
+            - Its longitude is greater than 180 degrees or less than -180 degrees.
+            
+            see https://en.wikipedia.org/wiki/Decimal_degrees
+        """;
+
+        if (!_isValidCoordinate(this.latitude1, this.longitude1))
+            throw new FormatException("Invalid coordinates at latitude1|longitude1\n$invalidDescription");
+
+        if (!_isValidCoordinate(this.latitude2, this.longitude2))
+            throw new FormatException("Invalid coordinates at latitude2|longitude2\n$invalidDescription");
     }
 
     double distance() {
@@ -48,7 +69,16 @@ class Haversine {
         return distance;
     }
 
-    double _radiansFromDegrees(final double degrees) {
-        return degrees * (PI / 180.0);
-    }
+    double _radiansFromDegrees(final double degrees) => degrees * (PI / 180.0);
+
+    /// A coordinate is considered invalid if it meets at least one of the following criteria:
+    ///
+    /// - Its latitude is greater than 90 degrees or less than -90 degrees.
+    ///- Its longitude is greater than 180 degrees or less than -180 degrees.
+    bool _isValidCoordinate(double latitude, longitude) => _isValidLatitude(latitude) && _isValidLongitude(longitude);
+    
+    /// A latitude is considered invalid if its is greater than 90 degrees or less than -90 degrees.
+    bool  _isValidLatitude(double latitudeInRadians) => !(latitudeInRadians < _radiansFromDegrees(-90.0) || latitudeInRadians > _radiansFromDegrees(90.0));
+    /// A longitude is considered invalid if its is greater than 180 degrees or less than -180 degrees.
+    bool _isValidLongitude(double longitudeInRadians) => !(longitudeInRadians < _radiansFromDegrees(-180.0) || longitudeInRadians > _radiansFromDegrees(180.0));
 }