+++ test/haversine_test.dart
@@ -61,4 +61,70 @@ void main() {
             expect(distance, lessThan(8));
         });
     });
+
+    group('Simple coordinate validation tests', () {
+        Haversine harvesine;
+        final lat1 = 41.139129;
+        final lon1 = 1.402244;
+
+        final lat2 = 41.139074;
+        final lon2 = 1.402315;
+
+        setUp(() {
+            harvesine = new Haversine.fromDegrees(
+                latitude1: lat1, longitude1: lon1, latitude2: lat2, longitude2: lon2);
+        });
+
+        test('Having valid coordinates should no throw', () {
+            expect(() => harvesine.distance(), returnsNormally);
+        });
+
+        test('Constructor fromDegrees Having invalid latitude at coordinates 1 should throw FormatException', () {
+            final lat1 = -122.4612387012154;
+            final lon1 = 1.402244;
+
+            final lat2 = 41.139074;
+            final lon2 = 1.402315;
+
+            expect(() => new Haversine.fromDegrees(
+                latitude1: lat1, longitude1: lon1, latitude2: lat2, longitude2: lon2),
+                throwsA(predicate((e) => e is FormatException && e.message.contains("latitude1|longitude1"))));
+        });
+
+        test('Constructor fromDegrees Having invalid latitude at coordinates 2 should throw FormatException', () {
+            final lat1 = 41.139074;
+            final lon1 = 1.402244;
+
+            final lat2 = -122.4612387012154;
+            final lon2 = 1.402315;
+
+            expect(() => new Haversine.fromDegrees(
+                latitude1: lat1, longitude1: lon1, latitude2: lat2, longitude2: lon2),
+                throwsA(predicate((e) => e is FormatException && e.message.contains("latitude2|longitude2"))));
+        });
+
+        test('Constructor fromDegrees Having invalid lontigute at coordinates 2 should throw FormatException', () {
+            final lat1 = 41.139129;
+            final lon1 = 1.402244;
+
+            final lat2 = 41.139074;
+            final lon2 = 270.0;
+
+            expect(() => new Haversine.fromDegrees(
+                latitude1: lat1, longitude1: lon1, latitude2: lat2, longitude2: lon2),
+                throwsA(predicate((e) => e is FormatException && e.message.contains("latitude2|longitude2"))));
+        });
+
+        test('Constructor fromRadians Having invalid lontigute at coordinates 2 should throw FormatException', () {
+            final lat1 = 0.718013252451;
+            final lon1 = 0.024473774716;
+
+            final lat2 = 0.71801229252;
+            final lon2 = 4.71239;
+
+            expect(() => new Haversine.fromRadians(
+                latitude1: lat1, longitude1: lon1, latitude2: lat2, longitude2: lon2),
+                throwsA(predicate((e) => e is FormatException && e.message.contains("latitude2|longitude2"))));
+        });
+    });
 }